Add support for authentication.
Controlling a squeezebox setup via mysqueezebox.com is now possible.

Authentication is performed by sending a username & a password to
mysqueezebox.com which replies with an authentication cookie that should
be included in all future requests. The authentication method has been
replicated from the one performed by the web remote control found on
http://www.mysqueezebox.com/index/Home.